feat: Move eudi utils to EUDIUtils.kt and draft message to display if not registered with EUDI.
