+++ common/build.gradle
@@ -136,6 +136,7 @@ dependencies {
     implementation "app.cash.sqldelight:android-driver:$sqldelight_version"
     implementation "app.cash.sqldelight:coroutines-extensions:$sqldelight_version"
     implementation 'com.google.zxing:core:3.4.1'
+    implementation("com.squareup.okhttp3:okhttp:4.12.0")
     testImplementation 'io.mockk:mockk-dsl:1.13.9'
 }
