+++ eurotoken/src/main/java/nl/tudelft/trustchain/eurotoken/ui/transfer/SendMoneyFragment.kt
@@ -5,13 +5,16 @@ import android.os.Bundle
 import android.view.View
 import android.widget.Toast
 import androidx.core.content.ContextCompat
+import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
+import kotlinx.coroutines.launch
 import nl.tudelft.ipv8.keyvault.defaultCryptoProvider
 import nl.tudelft.ipv8.util.hexToBytes
 import nl.tudelft.ipv8.util.toHex
 import nl.tudelft.trustchain.common.contacts.ContactStore
 import nl.tudelft.trustchain.common.eurotoken.TransactionRepository
 import nl.tudelft.trustchain.common.util.viewBinding
+import nl.tudelft.trustchain.common.util.EUDIUtils
 import nl.tudelft.trustchain.eurotoken.EuroTokenMainActivity
 import nl.tudelft.trustchain.eurotoken.R
 import nl.tudelft.trustchain.eurotoken.databinding.FragmentSendMoneyBinding
@@ -21,6 +24,10 @@ class SendMoneyFragment : EurotokenBaseFragment(R.layout.fragment_send_money) {
     private var addContact = false
 
     private val binding by viewBinding(FragmentSendMoneyBinding::bind)
+
+    private val eudiUtils by lazy {
+        EUDIUtils()
+    }
 
     private val ownPublicKey by lazy {
         defaultCryptoProvider.keyFromPublicBin(
@@ -93,17 +100,14 @@ class SendMoneyFragment : EurotokenBaseFragment(R.layout.fragment_send_money) {
         val trustScore = trustStore.getScore(publicKey.toByteArray())
         logger.info { "Trustscore: $trustScore" }
 
-        if (trustScore != null) {
-            if (trustScore >= TRUSTSCORE_AVERAGE_BOUNDARY) {
-                binding.trustScoreWarning.text =
-                    getString(R.string.send_money_trustscore_warning_high, trustScore)
-                binding.trustScoreWarning.setBackgroundColor(
-                    ContextCompat.getColor(
-                        requireContext(),
-                        R.color.android_green
-                    )
-                )
-            } else if (trustScore > TRUSTSCORE_LOW_BOUNDARY) {
+
+        lifecycleScope.launch {
+            if (!eudiUtils.verifyEudiToken()) {
+                /*
+TODO: Change contents here below. Check if user registered using EUDI, if so display trust message
+        // Check if user we want to send money is already registered on the chain.
+
+                 */
                 binding.trustScoreWarning.text =
                     getString(R.string.send_money_trustscore_warning_average, trustScore)
                 binding.trustScoreWarning.setBackgroundColor(
@@ -112,26 +116,46 @@ class SendMoneyFragment : EurotokenBaseFragment(R.layout.fragment_send_money) {
                         R.color.metallic_gold
                     )
                 )
+            } else if (trustScore != null) {
+                if (trustScore >= TRUSTSCORE_AVERAGE_BOUNDARY) {
+                    binding.trustScoreWarning.text =
+                        getString(R.string.send_money_trustscore_warning_high, trustScore)
+                    binding.trustScoreWarning.setBackgroundColor(
+                        ContextCompat.getColor(
+                            requireContext(),
+                            R.color.android_green
+                        )
+                    )
+                } else if (trustScore > TRUSTSCORE_LOW_BOUNDARY) {
+                    binding.trustScoreWarning.text =
+                        getString(R.string.send_money_trustscore_warning_average, trustScore)
+                    binding.trustScoreWarning.setBackgroundColor(
+                        ContextCompat.getColor(
+                            requireContext(),
+                            R.color.metallic_gold
+                        )
+                    )
+                } else {
+                    binding.trustScoreWarning.text =
+                        getString(R.string.send_money_trustscore_warning_low, trustScore)
+                    binding.trustScoreWarning.setBackgroundColor(
+                        ContextCompat.getColor(
+                            requireContext(),
+                            R.color.red
+                        )
+                    )
+                }
             } else {
                 binding.trustScoreWarning.text =
-                    getString(R.string.send_money_trustscore_warning_low, trustScore)
+                    getString(R.string.send_money_trustscore_warning_no_score)
                 binding.trustScoreWarning.setBackgroundColor(
                     ContextCompat.getColor(
                         requireContext(),
-                        R.color.red
+                        R.color.metallic_gold
                     )
                 )
+                binding.trustScoreWarning.visibility = View.VISIBLE
             }
-        } else {
-            binding.trustScoreWarning.text =
-                getString(R.string.send_money_trustscore_warning_no_score)
-            binding.trustScoreWarning.setBackgroundColor(
-                ContextCompat.getColor(
-                    requireContext(),
-                    R.color.metallic_gold
-                )
-            )
-            binding.trustScoreWarning.visibility = View.VISIBLE
         }
 
         binding.btnSend.setOnClickListener {
