+++ eurotoken/src/main/java/nl/tudelft/trustchain/eurotoken/ui/transfer/TransferFragment.kt
@@ -29,6 +29,7 @@ import nl.tudelft.ipv8.util.toHex
 import nl.tudelft.trustchain.common.contacts.ContactStore
 import nl.tudelft.trustchain.common.eurotoken.TransactionRepository
 import nl.tudelft.trustchain.common.util.QRCodeUtils
+import nl.tudelft.trustchain.common.util.EUDIUtils
 import nl.tudelft.trustchain.common.util.WebAuthnIdentityProviderOwner
 import nl.tudelft.trustchain.common.util.viewBinding
 import nl.tudelft.trustchain.eurotoken.EuroTokenMainActivity
@@ -36,19 +37,20 @@ import nl.tudelft.trustchain.eurotoken.R
 import nl.tudelft.trustchain.eurotoken.community.EuroTokenCommunity
 import nl.tudelft.trustchain.eurotoken.databinding.FragmentTransferEuroBinding
 import nl.tudelft.trustchain.eurotoken.ui.EurotokenBaseFragment
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
 import org.json.JSONException
 import org.json.JSONObject
 
-
+const val TOON_MSG = "ToonsStuff"
+const val EUROTOKEN_MSG = "EUROTOKEN"
 class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) {
     private val binding by viewBinding(FragmentTransferEuroBinding::bind)
 
     private val qrCodeUtils by lazy {
         QRCodeUtils(requireContext())
+    }
+
+    private val eudiUtils by lazy {
+        EUDIUtils()
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -180,24 +182,24 @@ class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) 
             lifecycleScope.launch {
                 val myPublicKey = transactionRepository.getGatewayPeer()?.publicKey?.keyToBin()
                     ?: throw Error("Could not find public key")
-                 val verifyResult = verifyEudiToken()
+                 val verifyResult = eudiUtils.verifyEudiToken()
                  if (verifyResult == false) {
-                     Log.d("ToonsStuff", "Failed to verify EUDI token")
+                     Log.d(TOON_MSG, "Failed to verify EUDI token")
                      Toast.makeText(requireActivity(), "Failed to verify EUDI token", Toast.LENGTH_LONG).show()
                      return@launch
                  }
                 val eudiToken = getEudiToken().toString()
-                Log.d("ToonsStuff", "EudiToken $eudiToken")
+                Log.d(TOON_MSG, "EudiToken $eudiToken")
 
                 val myIdentityProvider: WebAuthnIdentityProviderOwner =
                     (getIpv8().myPeer.identityProvider ?: throw Error("big problems bro")) as WebAuthnIdentityProviderOwner
-                Log.d("ToonsStuff", "Identity provider: $myIdentityProvider")
+                Log.d(TOON_MSG, "Identity provider: $myIdentityProvider")
                 // I don't like this ^ :(
                 // also fixing it will be a bit of a mess but oh well
 
                 myIdentityProvider.context = requireActivity()
                 val signedKey = myIdentityProvider.sign(eudiToken.toByteArray())
-                Log.d("ToonsStuff", "Signed token: $signedKey")
+                Log.d(TOON_MSG, "Signed token: $signedKey")
 
                 val transaction = mapOf("signed_EUDI_key" to signedKey?.toJsonString())
 
@@ -207,163 +209,37 @@ class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) 
                     myPublicKey
                 )
                 transactionRepository.trustChainCommunity.getPeers().forEach { peer ->
-                    Log.d("ToonsStuff", "Sending to peer: " + peer.address)
+                    Log.d(TOON_MSG, "Sending to peer: " + peer.address)
                     transactionRepository.trustChainCommunity.sendBlock(block, peer)
                 }
                 transactionRepository.trustChainCommunity.addListener(
                     "eurotoken_register",
                     object : BlockListener {
                         override fun onBlockReceived(block: TrustChainBlock) {
-                            Log.d("ToonsStuff", "blockReceived: ${block.blockId} ${block.transaction}")
+                            Log.d(TOON_MSG, "blockReceived: ${block.blockId} ${block.transaction}")
                         }
                     }
                 )
-                Log.d("ToonsStuff", "Size of db:  ${transactionRepository.trustChainCommunity.database.getAllBlocks().size}")
-                Log.d("ToonsStuff", transactionRepository.trustChainCommunity.getChainLength().toString())
+                Log.d(TOON_MSG, "Size of db:  ${transactionRepository.trustChainCommunity.database.getAllBlocks().size}")
+                Log.d(TOON_MSG, transactionRepository.trustChainCommunity.getChainLength().toString())
                 Toast.makeText(requireActivity(), "Registered on the ⛓\uFE0Fchain\uD83D\uDE80", Toast.LENGTH_LONG).show()
             }
         }
     }
 
     suspend fun getEudiToken(): JSONObject {
-        Log.d("ToonsStuff", "Opening EUDI app")
-
-        val content = """
-                    {
-                    "type": "vp_token",
-                    "presentation_definition": {
-                        "id": "1e7896b5-bbcc-4730-94b2-8232cfac2658",
-                        "input_descriptors": [
-                        {
-                            "id": "f290d465-3fff-4637-89f1-08f8606ccd7b",
-                            "name": "Person Identification Data (PID)",
-                            "purpose": "",
-                            "format": {
-                            "dc+sd-jwt": {
-                                "sd-jwt_alg_values": [
-                                "ES256",
-                                "ES384",
-                                "ES512"
-                                ],
-                                "kb-jwt_alg_values": [
-                                "RS256",
-                                "RS384",
-                                "RS512",
-                                "ES256",
-                                "ES384",
-                                "ES512"
-                                ]
-                            }
-                            },
-                            "constraints": {
-                            "fields": [
-                                {
-                                "path": [
-                                    "$.vct"
-                                ],
-                                "filter": {
-                                    "type": "string",
-                                    "const": "urn:eu.europa.ec.eudi:pid:1"
-                                }
-                                },
-                                {
-                                "path": [
-                                    "$.family_name"
-                                ],
-                                "intent_to_retain": false
-                                },
-                                {
-                                "path": [
-                                    "$.given_name"
-                                ],
-                                "intent_to_retain": false
-                                }
-                            ]
-                            }
-                        }
-                        ]
-                    },
-                    "nonce": "2418429c-f59f-4b48-99c1-4f4bfaff8116",
-                    "request_uri_method": "get"
-                    }
-                """.trimIndent()
-        val verifierData = makeApiCall("https://verifier-backend.eudiw.dev/ui/presentations", "POST", content) ?: JSONObject() // Unsafe, how to handle null case?
-        Log.d("ToonsStuff", "Found my cool URL: $verifierData")
-
-        val transactionId = verifierData.getString("transaction_id")
-        val clientId = verifierData.getString("client_id")
-        val requestUri = verifierData.getString("request_uri")
-        val requestUriMethod = verifierData.getString("request_uri_method")
-
-        val url = "eudi-openid4vp://?client_id=$clientId&request_uri=$requestUri&request_uri_method=$requestUriMethod"
+        val tokenPair = eudiUtils.getEudiToken()
+        val url = tokenPair.second
 
         val intent = Intent(Intent.ACTION_VIEW).apply {
             data = url.toUri()
             flags = Intent.FLAG_ACTIVITY_NEW_TASK
         }
 
-        Log.d("ToonsStuff", "Starting activity")
+        Log.d(TOON_MSG, "Starting activity")
         startActivity(intent)
 
-        val getWalletUrl = "https://verifier-backend.eudiw.dev/ui/presentations/$transactionId"
-        // Primitive handling of EUDIW stuff, is there no better way?
-        while (true) {
-            delay(1000)
-            val walletResult = makeApiCall(getWalletUrl, "GET", body = null)
-            if (walletResult != null) {
-                val vpTokenarray = walletResult.getJSONArray("vp_token")
-                val vpToken = vpTokenarray[0].toString()
-                Log.d("ToonsStuff", "Received VP token from thingy: $vpToken")
-                return walletResult
-            } else {
-                Log.d("ToonsStuff", "Failed to get wallet results")
-            }
-        }
-    }
-
-    suspend fun verifyEudiToken(token: JSONObject? = null): Boolean {
-        try {
-            Log.d("ToonsStuff", "Starting EUDI token verification")
-
-            val walletResult = token ?: getEudiToken()
-
-            val vpTokenArray = walletResult.getJSONArray("vp_token")
-            val vpTokenString = vpTokenArray.getString(0)
-
-            Log.d("ToonsStuff", "Extracted JWT: $vpTokenString")
-
-            val verifyRequestBody = JSONObject().apply {
-                put("sd_jwt_vc", vpTokenString)
-                put("nonce", "2418429c-f59f-4b48-99c1-4f4bfaff8116")
-            }.toString()
-
-            val verificationResponse = makeApiCall(
-                "https://verifier-backend.eudiw.dev/utilities/validations/sdJwtVc",
-                "POST",
-                verifyRequestBody
-            )
-
-            if (verificationResponse == null) {
-                Log.e("ToonsStuff", "Validation failed - null response")
-                return false
-            }
-
-            val isValid = verificationResponse.optBoolean("valid", false)
-
-            if (isValid) {
-                Log.d("ToonsStuff", "SD-JWT verified successfully by EUDI validator")
-                return true
-            } else {
-                val errorMessage = verificationResponse.optString("error", "Unknown validation error")
-                Log.e("ToonsStuff", "Token validation failed: $errorMessage")
-                return false
-            }
-
-        } catch (e: Exception) {
-            Log.e("ToonsStuff", "Error verifying token: ${e.message}")
-            e.printStackTrace()
-            return false
-        }
+        return tokenPair.first
     }
 
     /**
@@ -374,7 +250,7 @@ class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) 
         val itr = transactionRepository.trustChainCommunity.getPeers().listIterator()
         while (itr.hasNext()) {
             val cur: Peer = itr.next()
-            Log.d("EUROTOKEN", cur.key.pub().toString())
+            Log.d(EUROTOKEN_MSG, cur.key.pub().toString())
             if (cur.key.pub().toString() == pubKey) {
                 return cur
             }
@@ -472,7 +348,7 @@ class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) 
         }
 
         fun getAmount(amount: String): Long {
-            val regex = """[^\d]""".toRegex()
+            val regex = """\D""".toRegex()
             if (amount.isEmpty()) {
                 return 0L
             }
@@ -528,25 +404,5 @@ class TransferFragment : EurotokenBaseFragment(R.layout.fragment_transfer_euro) 
                 }
             )
         }
-    }
-
-    suspend fun makeApiCall(url: String, method: String, body: String?): JSONObject? = withContext(Dispatchers.IO) {
-        val request = Request.Builder()
-            .url(url)
-            .method(method, body?.toRequestBody("application/json; charset=utf-8".toMediaTypeOrNull()))
-            .build()
-        try {
-            OkHttpClient().newCall(request).execute().use { response ->
-                Log.d("ToonsStuff", "received status: ${response.code}")
-                Log.d("ToonsStuff", "received response: ${response.body?.toString()}")
-                val str = response.body?.string() // TODO: Unsafe, what to do in case of malformed body?
-
-                val data = str?.let { s -> JSONObject(s) }
-                return@withContext data
-            }
-        } catch (e: Exception) {
-            Log.d("ToonsStuff", "Exception during API call: $e")
-        }
-       return@withContext null
     }
 }
